Implement mutable builders and .build() method
Mutable builders are convenient because they allow the reducer to be
built inline with the handler declarations (see
https://github.com/dphilipson/typescript-fsa-reducers/issues/7). This
commit makes the builders mutable while also providing a .build() method
to produce non-mutable reducers. This also helps with the export problem
of impor[ting a reducer built with this library and ending up with an
unrecognized type ReducerBuilder error unless this library is also
imported.